Rewrote how the core URL system works. With a little bit of luck the collateral damage will be minimized and we can comfortably migrate the other apps.
